+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 ### Bugfixes
 
 - [#8091](https://github.com/influxdata/influxdb/issues/8091): Do not increment the continuous query statistic if no query is run.
+- [#8064](https://github.com/influxdata/influxdb/issues/8064): Forbid wildcards in binary expressions.
 
 ## v1.2.2 [2017-03-14]