+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - [#8091](https://github.com/influxdata/influxdb/issues/8091): Do not increment the continuous query statistic if no query is run.
 - [#8064](https://github.com/influxdata/influxdb/issues/8064): Forbid wildcards in binary expressions.
 - [#8148](https://github.com/influxdata/influxdb/issues/8148): Fix fill(linear) when multiple series exist and there are null values.
+- [#7995](https://github.com/influxdata/influxdb/issues/7995): Update liner dependency to handle docker exec.
 
 ## v1.2.2 [2017-03-14]